requirements.txt
einops
huggingface_hub
matplotlib
numpy==1.23.4
omegaconf
openai==1.3.5
opencv-python>=4.4.0.46
openpyxl
pandas>=1.5.3
pillow
portalocker
protobuf
pycocoevalcap
python-dotenv
requests
rich
seaborn
sentencepiece
sty
tabulate
tiktoken
timeout-decorator
torch==2.1.2
tqdm
transformers
typing_extensions==4.7.1
validators
visual_genome
xlsxwriter